[Optimize][PageConfig] Extend bit width for feature flag in PipelineSchedulerConfig
In the current implementation, pipelineSchedulerConfig is defined as a 64-bit long integer to represent a combination of all scheduling-related feature flags, with each feature flag occupying one bit. However, using a single bit cannot represent the undefined status in TernaryBoolean.

In this merge request (MR), the bits occupied by each feature flag are extended to two bits to represent three states: 1 for enable, 0 for undefined, and 2 for disable. This change ensures that the online business will not be disrupted.

issue:m-0

AutoSubmit:true