[Optimize] Change gn path to relative path.
Now there are lots of gn absolute path usage in our project, it brings the following problem:
Affecting the parent directory of lynx: In order to make it possible to extend lynx project in other projects, we need to use //lynx gn absolute path in lynx project, so we should execute gn command in the parent directory of lynx. As a result, buildroot and other third_party libraries should be placed in parent directory of lynx.

To solve the problem, we prefer to use relative path in lynx project, refering to the approach in skia and v8.
In this change, we add a script named gn_relative_path_converter.py to help converting //lynx to relative path. But we still need to do some adaptions:
Add rebase_path function for converting relative path to absolute path in template definition and variable definition in build_overrides, because the relative path is different with different gn target.