[Testing] Avoid the disturbance caused by printing the full log
The test may generate a large volume of logs during certain stages, but only the
last few lines may be relevant to the actual execution. Printing all logs when a
test fails can introduce disturbances in troubleshooting, and may also lead to
log anomalies or truncation due to performance issues with the logging system.

Change-Id: Ie27dd990cfdc03c759fb5516637ba20a7eaa62fe