[BugFix][List] Fix list content offset clamped when initializing anchor info.
If we clamp content offset to the scroll range when initializing anchor info, we will calculate a wrong range which may lead to some visible area of list is not filled by list-item.

In this commit, we add with_clamp parameter in function SetContentOffset() and not clamp content offset when initializing anchor info, and add e2e / unit tests.

issue: m-6713268157