[BugFix][Element] Fix render with reused element bundle not trigger animation as expected
In the current implementation, when an element tree is cloned, the parsed_style_map is also copied to the new element. During style resolution for the cloned element tree, no style changes are detected, so keyframe and transition animations are not triggered because there is no difference between the previous and current states. This results in abnormal animation behavior.

In this merge request, a new dirty flag called kDirtyCloned is added to cloned elements. When resolving styles for these cloned elements, the presence of the kDirtyCloned flag is checked to force the reapplication of animation-related CSS properties, ensuring animations are triggered correctly.

issue:m-6733242178

AutoSubmit: true

AutoLand: release/3.3